+++ angular-workspace/projects/ni/nimble-angular/rich-text/editor/testing/ng-package.json
@@ -1,5 +1,5 @@
 {
-    "$schema": "../../../../../../node_modules/ng-packagr/ng-package.schema.json",
+    "$schema": "../../../../../../../node_modules/ng-packagr/ng-package.schema.json",
     "lib": {
       "entryFile": "public-api.ts"
     }
