+++ packages/nimble-components/src/rich-text/viewer/index.ts
@@ -1,8 +1,13 @@
 import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
+import {
+    schema,
+    defaultMarkdownParser,
+    MarkdownParser
+} from 'prosemirror-markdown';
+import { DOMSerializer } from 'prosemirror-model';
 import { observable } from '@microsoft/fast-element';
 import { template } from './template';
 import { styles } from './styles';
-import { RichTextMarkdownParser } from '../models/markdown-parser';
 
 declare global {
     interface HTMLElementTagNameMap {
@@ -26,8 +31,14 @@ export class RichTextViewer extends FoundationElement {
      * @internal
      */
     public viewer!: HTMLDivElement;
+    private readonly markdownParser: MarkdownParser;
+    private readonly domSerializer: DOMSerializer;
 
-    private readonly markdownParser = new RichTextMarkdownParser();
+    public constructor() {
+        super();
+        this.domSerializer = DOMSerializer.fromSchema(schema);
+        this.markdownParser = this.initializeMarkdownParser();
+    }
 
     /**
      * @internal
@@ -46,11 +57,50 @@ export class RichTextViewer extends FoundationElement {
         }
     }
 
+    private initializeMarkdownParser(): MarkdownParser {
+        /**
+         * It configures the tokenizer of the default Markdown parser with the 'zero' preset.
+         * The 'zero' preset is a configuration with no rules enabled by default to selectively enable specific rules.
+         * https://github.com/markdown-it/markdown-it/blob/2b6cac25823af011ff3bc7628bc9b06e483c5a08/lib/presets/zero.js#L1
+         *
+         */
+        const zeroTokenizerConfiguration = defaultMarkdownParser.tokenizer.configure('zero');
+
+        // The detailed information of the supported rules were provided in the below CommonMark spec document.
+        // https://spec.commonmark.org/0.30/
+        const supportedTokenizerRules = zeroTokenizerConfiguration.enable([
+            'emphasis',
+            'list',
+            'autolink'
+        ]);
+
+        return new MarkdownParser(
+            schema,
+            supportedTokenizerRules,
+            defaultMarkdownParser.tokens
+        );
+    }
+
+    /**
+     *
+     * This function takes a markdown string, parses it using the ProseMirror MarkdownParser, serializes the parsed content into a
+     * DOM structure using a DOMSerializer, and returns the serialized result.
+     * If the markdown parser returns null, it will clear the viewer component by creating an empty document fragment.
+     */
+    private parseMarkdownToDOM(value: string): HTMLElement | DocumentFragment {
+        const parsedMarkdownContent = this.markdownParser.parse(value);
+        if (parsedMarkdownContent === null) {
+            return document.createDocumentFragment();
+        }
+
+        return this.domSerializer.serializeFragment(
+            parsedMarkdownContent.content
+        );
+    }
+
     private updateView(): void {
         if (this.markdown) {
-            const serializedContent = this.markdownParser.parseMarkdownToDOM(
-                this.markdown
-            );
+            const serializedContent = this.parseMarkdownToDOM(this.markdown);
             this.viewer.replaceChildren(serializedContent);
         } else {
             this.viewer.innerHTML = '';
